Switch to SDL2 and display a tringle on OSX.
Switch to SDL2 and finally got a colored triangle to draw to the screen.

Winit and SDL2 are complete nightmates when you try to use them with Filament.
Between linker errors, and the phantom issue of low framerates, it's been 3
days trying to get a triangle to draw to the screen.